[feat] : 패턴 모델 연결
